requirements.txt
greenlet==1.1.2
Pillow==9.2.0
SQLAlchemy==1.4.39
